requirements.txt
localStoragePy
opencv-python
customtkinter
pyinstaller
jellyfish
requests
Pillow
numpy
regex
ttkbootstrap